# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/operations": {
            "get": {
                "description": "retrieve the list of operations",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Operations"
                ],
                "summary": "Get the operations list",
                "operationId": "get-operations",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter parameter",
                        "name": "filter_param",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter value",
                        "name": "filter_value",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort field",
                        "name": "sort_field",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort order",
                        "name": "sort_order",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/repositories.PaginatedOperations"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorMessage"
                        }
                    }
                }
            },
            "post": {
                "description": "create a new operation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Operations"
                ],
                "summary": "Create a new operation",
                "operationId": "post-operation",
                "parameters": [
                    {
                        "description": "Operation object",
                        "name": "operation",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.OperationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.OperationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorMessage"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorMessage"
                        }
                    }
                }
            }
        },
...
        "/api/v1/operations/types": {
            "get": {
                "description": "retrieve the list of operations types",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OperationsTypes"
                ],
                "summary": "Get the operations types list",
                "operationId": "get-operations-types",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/operation.OperationType"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorMessage"
                        }
                    }
                }
            },
            "post": {
                "description": "create a new operation type",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OperationsTypes"
                ],
                "summary": "Create a new operation type",
                "operationId": "post-operation-type",
                "parameters": [
                    {
                        "description": "Operation Type",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.SaveOperationTypeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.Result"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorMessage"
                        }
                    }
                }
            },
            "patch": {
                "description": "update a operation type",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OperationsTypes"
                ],
                "summary": "Update a operation type",
                "operationId": "patch-operation-type",
                "parameters": [
                    {
                        "description": "Operation Type",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.SaveOperationTypeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.Result"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorMessage"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorMessage"
                        }
                    }
                }
            }
        },
        "/api/v1/operations/types/dashboard": {
            "get": {
                "description": "retrieve the list of operations types names",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OperationsTypes"
                ],
                "summary": "Get the operations types names list",
                "operationId": "get-operations-types-names",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorMessage"
                        }
                    }
                }
            }
        },
        "/api/v1/operations/types/{id}": {
            "get": {
                "description": "retrieve a operation type by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OperationsTypes"
                ],
                "summary": "Get a operation type",
                "operationId": "get-operation-type-by-id",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/operation.OperationType"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorMessage"
                        }
                    }
                }
            },
            "delete": {
                "description": "delete a operation type",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OperationsTypes"
                ],
                "summary": "Delete a operation type",
                "operationId": "delete-operation-type",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Operation Type ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.Result"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorMessage"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "operation.OperationType": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "repositories.Operation": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string"
                },
                "blockchain_status": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "domain": {
                    "type": "string"
                },
                "fireblocks_id": {
                    "type": "string"
                },
                "fireblocks_status": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "operator": {
                    "type": "string"
                },
                "transaction_hash": {
                    "type": "string"
                },
                "transaction_link": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "repositories.PaginatedOperations": {
            "type": "object",
            "properties": {
                "current_page": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/repositories.Operation"
                    }
                },
                "next_page": {
                    "type": "integer"
                },
                "previous_page": {
                    "type": "integer"
                },
                "total_count": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
        "types.ErrorMessage": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "types.OperationRequest": {
            "type": "object",
            "required": [
                "amount",
                "blockchain_id",
                "domain",
                "operator",
                "token_id",
                "type"
            ],
            "properties": {
                "amount": {
                    "type": "string",
                    "example": "2.75"
                },
                "blockchain_id": {
                    "type": "string",
                    "example": "66f6fe7eccc6398d39e981f9"
                },
                "domain": {
                    "type": "string",
                    "enum": [
                        "GET-BRAZA",
                        "BRAZA-ON",
                        "BRAZA-DESK"
                    ],
                    "example": "GET-BRAZA"
                },
                "operator": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                },
                "token_id": {
                    "type": "string",
                    "example": "66f74acbba6b56108cb3e80a"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "MINT",
                        "BURN"
                    ],
                    "example": "MINT,BURN"
                }
            }
        },
        "types.OperationResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "operation was accepted to be processed on blockchain"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "types.Result": {
            "type": "object",
            "properties": {
                "result": {
                    "type": "string"
                }
            }
        },
...
        "types.SaveOperationTypeRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "example": "TYPE-NAME"
                }
            }
        },
...
    }
}